asim: move queues into separate pkg
This patch separates out the replicate queue and split queue into
separate files. This patch also moves the respective queues and the base
queue into a separate package within asim.

Release note: None
