+++ pkg/BUILD.bazel
@@ -172,6 +172,7 @@ ALL_TESTS = [
     "//pkg/kv/kvserver/allocator/allocatorimpl:allocatorimpl_test",
     "//pkg/kv/kvserver/allocator/storepool:storepool_test",
     "//pkg/kv/kvserver/apply:apply_test",
+    "//pkg/kv/kvserver/asim/queue:queue_test",
     "//pkg/kv/kvserver/asim/state:state_test",
     "//pkg/kv/kvserver/asim/workload:workload_test",
     "//pkg/kv/kvserver/asim:asim_test",
@@ -1063,6 +1064,8 @@ GO_TARGETS = [
     "//pkg/kv/kvserver/apply:apply",
     "//pkg/kv/kvserver/apply:apply_test",
     "//pkg/kv/kvserver/asim/config:config",
+    "//pkg/kv/kvserver/asim/queue:queue",
+    "//pkg/kv/kvserver/asim/queue:queue_test",
     "//pkg/kv/kvserver/asim/state:state",
     "//pkg/kv/kvserver/asim/state:state_test",
     "//pkg/kv/kvserver/asim/workload:workload",
@@ -2317,6 +2320,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/kv/kvserver/apply:get_x_data",
     "//pkg/kv/kvserver/asim:get_x_data",
     "//pkg/kv/kvserver/asim/config:get_x_data",
+    "//pkg/kv/kvserver/asim/queue:get_x_data",
     "//pkg/kv/kvserver/asim/state:get_x_data",
     "//pkg/kv/kvserver/asim/workload:get_x_data",
     "//pkg/kv/kvserver/batcheval:get_x_data",
